requirements.txt
crewai
crewai[tools]
python-dotenv
pydantic
google-generativeai
rq